Revert "zt/fix loading"
