+++ lib/atrium/client.rb
@@ -1,0 +1,42 @@
+module Atrium
+  class Client
+    attr_accessor :mx_api_key, :mx_client_id
+
+    def initialize(api_key = nil, client_id = nil)
+      @mx_api_key = api_key
+      @mx_client_id = client_id
+    end
+
+    def make_request(method, endpoint, body = {}, headers = {})
+      headers = default_headers.merge(headers)
+      url = "#{::Atrium::BASE_URL}/#{endpoint}"
+      response = http_client.public_send(method, url, ::JSON.dump(body), headers)
+
+      handle_response(response)
+    end
+
+    def http_client
+      @http_client ||= ::HTTPClient.new
+    end
+
+  private
+
+    def default_headers
+      {
+        "Accept" => "application/vnd.mx.atrium.v1+json",
+        "Content-Type" => "application/json",
+        "MX-API-KEY" => mx_api_key,
+        "MX-CLIENT-ID" => mx_client_id
+      }
+    end
+
+    def handle_response(response)
+      # Handle 200-206 responses as acceptable
+      unless response.status.between?(200, 206)
+        fail ::Atrium::Error, "#{response.status}: #{response.body}"
+      end
+
+      ::JSON.parse(response.body) unless response.body.empty?
+    end
+  end
+end
