+++ lib/atrium/error.rb
@@ -1,0 +1,4 @@
+module Atrium
+  class Error < StandardError
+  end
+end
