+++ lib/atrium/version.rb
@@ -1,0 +1,3 @@
+module Atrium
+  VERSION = "0.2.3"
+end
