--- lib/mx/atrium/credential.rb
@@ -1,13 +1,0 @@
-module MX
-  module Atrium
-    class Credential
-      include ::ActiveAttr::Model
-
-      attribute :field_name
-      attribute :guid
-      attribute :label
-      attribute :type
-      attribute :value
-    end
-  end
-end
