--- lib/mx/atrium/error.rb
@@ -1,6 +1,0 @@
-module MX
-  module Atrium
-    class Error < StandardError
-    end
-  end
-end
