--- lib/mx/atrium/version.rb
@@ -1,5 +1,0 @@
-module MX
-  module Atrium
-    VERSION = "0.2.3"
-  end
-end
